% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lTab.R
\name{lTab}
\alias{lTab}
\title{Produces a latex longtable}
\usage{
lTab(x, label = NULL, cap.head = NULL, cap.fhead = NULL,
  cap.foot = NULL, cap.lfoot = NULL, booktabs = .op("kLat.lTab.booktabs",
  "kLat.booktabs", FALSE), toprule = .book("kLat.toprule", booktabs,
  "\\\\toprule", "\\\\hline"), botrule = .book("kLat.botrule", booktabs,
  "\\\\bottomrule", "\\\\hline"), midrule = .book("kLat.midrule", booktabs,
  "\\\\midrule", "\\\\hline"), rowsep = .op("kLat.lTab.rowsep", "kLat.rowsep",
  ""), rows = .op("kLat.lTab.rows", "kLat.rows", FASLE),
  colsep = .op("kLat.lTab.colsep", "kLat.colsep", ""), coldef = .coldef(x,
  colsep), align = .op("kLat.lTab.align", "kLat.align", "center"),
  header = .header(x, rows), envir = getOption("kLat.lTab.envir",
  "longtable"), fsthead = NULL, foot = botrule, lastfoot = NULL)
}
\arguments{
\item{x}{a data.frame or matrix to form the base of the table}

\item{cap.head,cap.fhead,cap.foot,cap.lfoot}{paces the caption in the head,
  firsthead, foot, or lastfoot respectively. It is important not to set a
  caption in an otherwise NULL section or strange bugs can occur. It is
  accaptable if the setting was set by default as in head and foot. Consult the longtable
  documentation for a more detailed explanation of these options.}

\item{booktabs}{logical value, if not set will use value of
  kLat.(xTab|sTab|lTab).booktabs, if not set will use value of kLat.booktabs,
  if not set defaults to false. When TRUE toprule defaults to '\\toprule',
  midrule to '\\midrule', and botrule to '\\bottomrule', when FALSE those
  values all default to '\\hline'. Has no effect when toprule, midrule, and
  botrule are individually set.}

\item{toprule}{sets the value for the top rule, if not set will be
  determined by the value of booktabs}

\item{botrule}{sets the value for the bottom rule, if not set will
  be determined by the value of booktabs}

\item{midrule}{sets the value for the mid rule, if not set will
  be determined by the value of booktabs}

\item{rowsep}{the separaotr to be used between rows (i.e. '\\hline'),
  defaults to ''}

\item{rows}{include rownames in table, by default the column names
  will be an empty string}

\item{colsep}{separaotr to be used between columns (i.e. '|'), defaults to
  '', if coldef is set this value is ignored and the separators must be
  specificed in the coldef}

\item{coldef}{sets column definition i.e. \\begin\{tabular\}\{'align'\},
  if not set defaults to numeric = right, character = left}

\item{header}{sets the value for the table header, defaults to the column
  names; if you set this be sure to end with '\\\\\\\\'}

\item{fsthead}{header on first page of table only; defaults to header; if you
  set this, you are responsible for setting any \\hline, \\toprule, or \\midrule
  lines}

\item{foot}{sets value of the table footer, defaults to the value of
  bottomrule}

\item{lastfoot}{footer on last page of table only}

\item{caption}{the caption for the table, unlinke xTab and sTab, there is no
  caption.top or captio.bottom option in longtable}
}
\description{
Produces a latex longtable
}
\examples{
lTab(mtcars)
}

