% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitLatex.R
\docType{package}
\name{knitLatex}
\alias{knitLatex}
\alias{knitLatex-package}
\title{Latex: helpers for knitr}
\description{
This package provides three main functions to assist in using knitr with
latex:
\itemize{
\item xTab: which createsw a basic latex table.
\item lTab: which creates a longtable.
\item sTab: creates a supertabular environment.
}
}

