% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lTab.R
\name{lTab}
\alias{lTab}
\title{Produces a latex longtable}
\usage{
lTab(x, label = NULL, caption = NULL,
  booktabs = getOption("knitLat.booktabs", FALSE),
  toprule = .book("knitLat.toprule", booktabs, "\\\\toprule", "\\\\hline"),
  botrule = .book("knitLat.botrule", booktabs, "\\\\bottomrule", "\\\\hline"),
  midrule = .book("knitLat.midrule", booktabs, "\\\\midrule", "\\\\hline"),
  colsep = "", align = .align(x, colsep), rowsep = "",
  rows = getOption("knitLat.rows", FALSE),
  position = getOption("knitLat.lTab.position", "c"), header = .header(x,
  rows), float = getOption("knitLat.lTab.float", "center"),
  envir = getOption("knitLat.lTab.envir", "longtable"), fsthead = NULL,
  foot = botrule, lastfoot = NULL)
}
\arguments{
\item{x}{a data.frame or matrix to form the base of the table}

\item{caption}{the caption for the table, unlinke xTab and sTab, there is no
  caption.top or captio.bottom option in longtable}

\item{booktabs}{if set to true will use booktabs' \\toprule, \\midrule, and
  \\bottomrule, if not set will use value of knitLat.booktabs, if that is
  not set defaults to false}

\item{toprule}{sets the value for the top rule, if not set will be
  determined by the value of booktabs}

\item{botrule}{sets the value for the bottom rule, if not set will
  be determined by the value of booktabs}

\item{midrule}{sets the value for the mid rule, if not set will
  be determined by the value of booktabs}

\item{align}{sets column alignment i.e. \\begin\{tabular\}\{'align'\},
  if not set defaults to numeric = right, character = left}

\item{rows}{include rownames in table, by default the column name
  wil be an empty string}

\item{position}{sets the position of the table i.e.
  \\begin\{table\}['position']}

\item{header}{sets the value for the table header, defaults to the column
  names; if you set this be sure to end with '\\\\\\\\'}

\item{fsthead}{header on first page of table only; defaults to header; if you
  set this, you are responsible for setting any \\hline, \\toprule, or \\midrule
  lines}

\item{foot}{sets value of the table footer, defaults to the value of
  bottomrule}

\item{lastfoot}{footer on last page of table only}
}
\description{
Produces a latex longtable
}
\examples{
lTab(mtcars)
}

