% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lTab.R
\name{lTab}
\alias{lTab}
\title{Produces a latex longtable}
\usage{
lTab(x, label = NULL, caption = NULL, booktabs = .op("kLat.lTab.booktabs",
  "kLat.booktabs", FALSE), toprule = .book("kLat.toprule", booktabs,
  "\\\\toprule", "\\\\hline"), botrule = .book("kLat.botrule", booktabs,
  "\\\\bottomrule", "\\\\hline"), midrule = .book("kLat.midrule", booktabs,
  "\\\\midrule", "\\\\hline"), colsep = .op("kLat.lTab.colsep", "kLat.colsep",
  ""), align = .align(x, colsep), rowsep = .op("kLat.lTab.rowsep",
  "kLat.rowsep", ""), rows = .op("kLat.lTab.rows", "kLat.rows", FASLE),
  float = .op("kLat.lTab.float", "kLat.float", "center"),
  position = getOption("kLat.lTab.position", "c"), header = .header(x,
  rows), envir = getOption("kLat.lTab.envir", "longtable"), fsthead = NULL,
  foot = botrule, lastfoot = NULL)
}
\arguments{
\item{x}{a data.frame or matrix to form the base of the table}

\item{caption}{the caption for the table, unlinke xTab and sTab, there is no
  caption.top or captio.bottom option in longtable}

\item{booktabs}{if set to true will use booktabs' \\toprule, \\midrule, and
  \\bottomrule, if not set will use value of knitLat.booktabs, if that is
  not set defaults to false}

\item{toprule}{sets the value for the top rule, if not set will be
  determined by the value of booktabs}

\item{botrule}{sets the value for the bottom rule, if not set will
  be determined by the value of booktabs}

\item{midrule}{sets the value for the mid rule, if not set will
  be determined by the value of booktabs}

\item{align}{sets column alignment i.e. \\begin\{tabular\}\{'align'\},
  if not set defaults to numeric = right, character = left}

\item{rows}{include rownames in table, by default the column name
  wil be an empty string}

\item{position}{sets the position of the table i.e.
  \\begin\{table\}['position']}

\item{header}{sets the value for the table header, defaults to the column
  names; if you set this be sure to end with '\\\\\\\\'}

\item{fsthead}{header on first page of table only; defaults to header; if you
  set this, you are responsible for setting any \\hline, \\toprule, or \\midrule
  lines}

\item{foot}{sets value of the table footer, defaults to the value of
  bottomrule}

\item{lastfoot}{footer on last page of table only}
}
\description{
Produces a latex longtable
}
\examples{
lTab(mtcars)
}

