% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sTab.R
\name{sTab}
\alias{sTab}
\title{Produces a latex supertabular environment}
\usage{
sTab(x, label = NULL, cap.top = NULL, cap.bot = NULL,
  booktabs = getOption("knitLat.booktabs", FALSE),
  toprule = .book("knitLat.toprule", booktabs, "\\\\toprule", "\\\\hline"),
  botrule = .book("knitLat.botrule", booktabs, "\\\\bottomrule", "\\\\hline"),
  midrule = .book("knitLat.midrule", booktabs, "\\\\midrule", "\\\\hline"),
  colsep = "", align = .align(x, colsep), rowsep = "",
  rows = getOption("knitLat.rows", FALSE),
  float = getOption("knitLat.sTab.float", "center"),
  envir = getOption("knitLat.sTab.envir", "supertabular"),
  header = .header(x, rows), fsthead = NULL, tabletail = botrule,
  lasttail = NULL)
}
\arguments{
\item{x}{a data.frame or matrix to form the base of the table}

\item{cap.top}{sets the caption command placing it at the top
  of the table}

\item{cap.bot}{sets the caption command placing it at the
  bottom of the table}

\item{booktabs}{if set to true will use booktabs' \\toprule, \\midrule, and
  \\bottomrule, if not set will use value of knitLat.booktabs, if that is
  not set defaults to false}

\item{toprule}{sets the value for the top rule, if not set will be
  determined by the value of booktabs}

\item{botrule}{sets the value for the bottom rule, if not set will
  be determined by the value of booktabs}

\item{midrule}{sets the value for the mid rule, if not set will
  be determined by the value of booktabs}

\item{align}{sets column alignment i.e. \\begin\{tabular\}\{'align'\},
  if not set defaults to numeric = right, character = left}

\item{rows}{include rownames in table, by default the column name
  wil be an empty string}

\item{header}{sets the value for the table header, defaults to the column
  names; if you set this be sure to end with '\\\\\\\\'}

\item{fsthead}{header on first page of table only}

\item{lasttail}{footer on last page of table only}

\item{head}{header to appear at the top of every page of table}

\item{tail}{footer on bottom of every page of table}
}
\description{
Produces a latex supertabular environment
}
\examples{
sTab(mtcars)
sTab(mtcars,
  cap.top = 'my super table',
  booktabs = TRUE,
  rows = TRUE)
}

