% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xTab.R
\name{xTab}
\alias{xTab}
\title{Produces a latex table}
\usage{
xTab(x, label = NULL, cap.top = NULL, cap.bot = NULL,
  position = getOption("knitLat.position", "ht"),
  booktabs = getOption("knitLat.booktabs", FALSE),
  toprule = .book("knitLat.toprule", booktabs, "\\\\toprule", "\\\\hline"),
  botrule = .book("knitLat.botrule", booktabs, "\\\\bottomrule", "\\\\hline"),
  midrule = .book("knitLat.midrule", booktabs, "\\\\midrule", "\\\\hline"),
  float = "center", envir = "tabular", colsep = "", rowsep = "",
  align = .align(x, colsep), rows = getOption("knitLat.rows", FALSE),
  header = .header(x, rows), foot = botrule)
}
\arguments{
\item{x}{a data.frame or matrix to form the base of the table}

\item{cap.top}{sets the caption command placing it at the top
  of the table}

\item{cap.bot}{sets the caption command placing it at the
  bottom of the table}

\item{position}{sets the position of the table i.e.
  \\begin\{table\}['position']}

\item{booktabs}{if set to true will use booktabs' \\toprule, \\midrule, and
  \\bottomrule, if not set will use value of knitLat.booktabs, if that is
  not set defaults to false}

\item{toprule}{sets the value for the top rule, if not set will be
  determined by the value of booktabs}

\item{botrule}{sets the value for the bottom rule, if not set will
  be determined by the value of booktabs}

\item{midrule}{sets the value for the mid rule, if not set will
  be determined by the value of booktabs}

\item{align}{sets column alignment i.e. \\begin\{tabular\}\{'align'\},
  if not set defaults to numeric = right, character = left}

\item{rows}{include rownames in table, by default the column name
  wil be an empty string}

\item{header}{sets the value for the table header, defaults to the column
  names; if you set this be sure to end with '\\\\\\\\'}

\item{foot}{sets value of the table footer, defaults to the value of
  bottomrule}
}
\description{
Produces a latex table
}
\examples{
xTab(mtcats)
xTab(mtcats, label='my table', caption.top='tab:mytable', booktabs=TRUE)
xTab(mtcats, head='col1 & col2 & \\\\eta\\\\\\\\')
}

